requirements.txt
Flask==2.0.2
Flask-WTF==0.15.1
gunicorn==20.1.0
PyJWT==2.3.0
Werkzeug==2.0.2
